requirements.txt
aniso8601==8.1.1
attrs==20.3.0
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Dance==3.2.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
pyrsistent==0.17.3
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
URLObject==2.4.3
Werkzeug==1.0.1
